feat: Add local time and date to all displays (Fixes #44)
Added a reusable `Clock` component to display the local time with a smaller date beneath it. This component was added to `src/app/kioskdisplay/page.tsx` and replaced the old time display in `src/app/kioskdisplay/certifications/page.tsx`.
